+++ test/src/core/escape/parser_test.mocks.dart
@@ -868,6 +868,14 @@ class MockEscapeHandler extends _i1.Mock implements _i2.EscapeHandler {
         returnValueForMissingStub: null,
       );
   @override
+  void setCurrentDirectory(String? uri) => super.noSuchMethod(
+        Invocation.method(
+          #setCurrentDirectory,
+          [uri],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
   void unknownOSC(String? code, List<String>? args) => super.noSuchMethod(
         Invocation.method(
           #unknownOSC,